Get the layout worker communicating back to the graph
